+++ tests/test_command.py
@@ -1,10 +1,13 @@
 from typing import Never
 
 import pytest
+import re
 
 from powercli import exceptions
 from powercli.args import Flag, Positional
 from powercli.command import Command
+from powercli.deprecation import Deprecation
+from powercli.typedefs import Context
 from powercli.utils import one_of, static
 
 
@@ -332,3 +335,34 @@ def test_subcommand() -> None:
     assert subargs is not None
     assert subargs.command.name == "foo"
     assert subargs.value_of_flag("f") == [12]
+
+
+def test_deprecation_with_bool() -> None:
+    cmd: Command[int, None]
+    cmd = Command()
+    cmd.flag(identifier="f", short="f", values=[("INT", int)])
+    cmd.flag(identifier="g", short="g", values=[("INT", int)], deprecation=lambda ctx: ctx.value_of("f") == [1])
+    with pytest.deprecated_call():
+        cmd.parse_args(["-f", "1", "-g", "2"])
+    args = cmd.parse_args(["-f", "2", "-g", "2"])
+    assert args.value_of_flag("f") == [2]
+    assert args.value_of_flag("g") == [2]
+
+def test_deprecation_with_object() -> None:
+    deprecation = Deprecation("use -h instead when -f is 1", since="1.0")
+    
+    # only for type checking
+    assert deprecation.message is not None
+    assert deprecation.since is not None
+
+    def f_deprecation(ctx: Context[int, None]) -> Deprecation | None:
+        if ctx.value_of("f") == [1]:
+            return deprecation
+        return None
+
+    cmd: Command[int, None]
+    cmd = Command()
+    cmd.flag(identifier="f", short="f", values=[("INT", int)])
+    cmd.flag(identifier="g", short="g", values=[("INT", int)], deprecation=f_deprecation)
+    with pytest.deprecated_call(match=f"{re.escape(deprecation.since)}: {re.escape(deprecation.message)}"):
+        cmd.parse_args(["-f", "1", "-g", "2"])