+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support for positionals with variable amount of values (#1).
+- Implementation of `powercli.args.Flag.deprecation` to deprecate flags (#2).
 - `powercli.utils.member_of` for easy integration of string enums.
 
 ### Changed